ReactDjango/backend/requirements.txt
asgiref
bson
django
djongo
mongo
pip
pymongo
pytz
setuptools
sqlparse==0.2.4
dnspython
djangorestframework
coverage
django-cors-headers
